+++ NEWS.md
@@ -2,7 +2,7 @@
 
 # development version
 
-- Add Perlstein 2026 extended-release injectable olanzapine TV-44749 and oral olanzapine ([doi:10.1002/jcph.70144](https://doi.org/10.1002/jcph.70144)) -- healthy adults and adults with schizophrenia or schizoaffective disorder.
+- Add Echterhof 2026 Pseudomonas bacteriophage whole-body PBPK ([doi:10.1128/aac.01506-25](https://doi.org/10.1128/aac.01506-25)) -- three models: CD-1 mice (the fitted species), plus rat and human interspecies projections.
 
 - Add Siebinga 2023 [68Ga]Ga- and [177Lu]Lu-HA-DOTATATE ([doi:10.1186/s40658-023-00565-4](https://doi.org/10.1186/s40658-023-00565-4)) - adults with neuroendocrine tumors.